+++ packages/react/src/tags-input/tags-input-field.tsx
@@ -3,13 +3,13 @@ import { forwardRef, type ComponentPropsWithoutRef } from 'react'
 import { ark } from '../factory'
 import { useTagsInputContext } from './tags-input-context'
 
-export type TagsInputFieldProps = ComponentPropsWithoutRef<typeof ark.input>
+export type TagsInputInputProps = ComponentPropsWithoutRef<typeof ark.input>
 
-export const TagsInputField = forwardRef<HTMLInputElement, TagsInputFieldProps>((props, ref) => {
+export const TagsInputInput = forwardRef<HTMLInputElement, TagsInputInputProps>((props, ref) => {
   const { inputProps } = useTagsInputContext()
   const mergedProps = mergeProps(inputProps, props)
 
   return <ark.input {...mergedProps} ref={ref} />
 })
 
-TagsInputField.displayName = 'TagsInputField'
+TagsInputInput.displayName = 'TagsInputInput'