+++ packages/solid/src/tags-input/tags-input-field.tsx
@@ -2,9 +2,9 @@ import { mergeProps } from '@zag-js/solid'
 import { ark, type HTMLArkProps } from '../factory'
 import { useTagsInputContext } from './tags-input-context'
 
-export type TagsInputFieldProps = HTMLArkProps<'input'>
+export type TagsInputInputProps = HTMLArkProps<'input'>
 
-export const TagsInputField = (props: TagsInputFieldProps) => {
+export const TagsInputInput = (props: TagsInputInputProps) => {
   const api = useTagsInputContext()
   const inputProps = mergeProps(() => api().inputProps, props)
   return <ark.input {...inputProps} />